+++ packages/expo-ui/android/src/main/java/expo/modules/ui/SliderView.kt
@@ -11,6 +11,7 @@ import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.ui.Modifier
 import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.records.Field
 import expo.modules.kotlin.records.Record
@@ -47,7 +48,7 @@ class SliderView(context: Context, appContext: AppContext) :
   private val onValueChanged by EventDispatcher()
 
   @Composable
-  override fun Content() {
+  override fun Content(modifier: Modifier) {
     val (value) = props.value
     val (min) = props.min
     val (max) = props.max